fix pod eviction storage
